requirements.txt
asgiref==3.7.2
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
Django==5.0.2
django-cors-headers==4.3.1
django-filter==23.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
Faker==24.2.0
h11==0.14.0
idna==3.6
Markdown==3.5.2
MouseInfo==0.1.3
numpy==1.26.4
opencv-python==4.9.0.80
outcome==1.3.0.post0
pillow==10.2.0
psycopg2==2.9.9
PyAutoGUI==0.9.54
pycparser==2.21
PyGetWindow==0.0.9
PyJWT==2.8.0
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytweening==1.0.7
pytz==2024.1
selenium==4.18.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.4.4
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
wsproto==1.2.0
